+++ CHANGELOG.md
@@ -6,3 +6,5 @@
   - Created new resource for replacing text in text files.
 - MSFT_KeyValuePairFile:
   - Created new resource for setting key value pairs in text files.
+- MSFT_IniSettingsFile:
+  - Created new resource for setting Windows INI file settings.